+++ code-examples/js/preview.js
@@ -1,3 +1,7 @@
+let renderedPrimaries = [];
+let renderedAlternates = [];
+let textNodes = [];
+
 const status = (message, type) => {
   const status = document.getElementById('hyper-init-status');
   status.getElementsByTagName('p')[0].innerHTML = message;
@@ -23,7 +27,7 @@ const loadMainJs = (server) =>
       () =>
         reject(
           new Error(
-            'Failed to load the main HyperIntelligence SDK JavaScript bundle.',
+            'Failed to load the main Hyper SDK JavaScript bundle.',
           ),
         ),
       5000,
@@ -35,12 +39,23 @@ const init = async () => {
   const server = s.server();
   await demo.validateServerURL(server);
   await loadMainJs(server);
-
+  const keywordInput = s.keywordInput();
   let authMode = s.authMode();
-  if (authMode.startsWith('mstrHyper.'))
+  if (authMode.startsWith('mstrHyper.')) {
     authMode = mstrHyper.AUTH_MODES[authMode.split('.').pop()];
+  }
+  const authModeRequiresPassword = [
+    mstrHyper.AUTH_MODES.STANDARD,
+    mstrHyper.AUTH_MODES.LDAP,
+  ].includes(authMode);
 
-  const onSessionErrorText = s.onSessionError();
+  const onSessionErrorText = s
+    .onSessionError()
+    .replace(/(var|let|const)\s+onSessionError\s*=/, 'window.onSessionError =')
+    .replace(
+      /function\s+onSessionError\s*\(/,
+      'window.onSessionError = function (',
+    );
   try {
     eval(onSessionErrorText);
   } catch (error) {
@@ -61,17 +76,31 @@ const init = async () => {
     }
   }
 
+  //Example onSearch callback which prints searchResults and searchId for each chunk
+  const onSearch = (searchResults, searchId) => {
+    console.log('This is an example of a custom onSearch callback.')
+    console.log('searchResults :', searchResults, `searchId :`, searchId);
+  };
+
+  //Example onSort callback which prints the sanitizedKeywords, sortingMap, and searchId once all search chunks are completed
+  const onSort = (sanitizedKeywords, sortingMap, searchId) => {
+    console.log('This is an example of a custom onSort callback.')
+    console.log('sanitizedKeywords :', sanitizedKeywords
+    , 'sortingMap :', sortingMap
+    , 'searchId :', searchId);
+  };
+
   const authToken = s.authToken() || null;
   const username = s.username() || null;
   const password = s.password() || (username ? '' : null);
-  return mstrHyper.start({
+  let hyperObject = await mstrHyper.start({
     server,
     auth: {
-      authMode: authMode,
-      username: authToken ? null : username,
-      password: authToken ? null : password,
+      authMode,
+      username: authToken || !authModeRequiresPassword ? null : username,
+      password: authToken || !authModeRequiresPassword ? null : password,
       authToken,
-      onSessionError: onSessionErrorText ? onSessionError : null,
+      onSessionError: onSessionErrorText ? window.onSessionError : null,
     },
     cards: cards ? hyperCards : null,
     logLevel: s.logLevel() || 'error',
@@ -79,8 +108,110 @@ const init = async () => {
       type: s.highlightType() || 'insertion',
       highlightIframes: s.highlightIframes() !== 'false',
     },
+    searchEnabled: true,
+    searching: {onSort, onSearch},
   });
+  // await mstrHyper.enableCards();
+  if(keywordInput.length > 0) {
+    await searchKeyword(keywordInput);
+  }
+  return hyperObject;
 };
+
+function addCountElement(count, parentNode){
+  let newDiv = document.createElement('div');
+  const newContent = document.createTextNode(`${count} card(s) not shown`);
+  newDiv.appendChild(newContent);
+  newDiv.style.backgroundColor = '#FFCC00';
+  newDiv.style.fontFamily = 'Robtoto';
+  newDiv.style.color = 'white';
+  newDiv.style.textAlign = 'center';
+  parentNode.insertBefore(newDiv, null);
+  return {node: newDiv, result: null};
+}
+
+async function addElement (result, parentNode) {
+  // create a new div element
+  let newDiv = document.createElement('div');
+  parentNode.insertBefore(newDiv, null);
+  await mstrHyper.showCard({elementId: result.ref, cardUID: result.cardSetId, nodeToRenderTo: newDiv});
+  return {node: newDiv, result};
+}
+
+function addSectionTextElement(sectionText, parentNode){
+  let newDiv = document.createElement('div');
+  const newContent = document.createTextNode(sectionText);
+  newDiv.appendChild(newContent);
+  newDiv.style.backgroundColor = '#24a0ed';
+  newDiv.style.fontFamily = 'Robtoto';
+  newDiv.style.color = 'white';
+  newDiv.style.textAlign = 'center';
+  parentNode.insertBefore(newDiv, null);
+  return {node: newDiv, result: null};
+}
+
+async function handleResultList(resultList, storageList, sectionText){
+  let IFrameNode = document.getElementById('IFrame');
+  let count = 0;
+  if(resultList.length > 0){
+    textNodes.push(addSectionTextElement(sectionText, IFrameNode));
+  }
+  for(result of resultList){
+    if(count == 5) {
+      textNodes.push(addCountElement(resultList.length - 5, IFrameNode));
+      break;
+    }
+    storageList.push(await addElement(result, IFrameNode));
+    count++;
+  }
+  return count > 0;
+}
+
+const searchKeyword = async(searchTerm = document.getElementById("searchTerm").value) => {
+  hideCard();
+  renderedPrimaries = [];
+  renderedPrimaries = [];
+  textNodes = [];
+  console.log("Searching: ", searchTerm)
+  let r = await mstrHyper.searchKeyword(searchTerm);
+  let results = await mstrHyper.mergeSearchResults(r.searchResults);
+  console.log("Searching results ", results);
+  let numRendered = 0;
+  numRendered += await handleResultList(results.primaryResults, renderedPrimaries, 'Primary Results');
+  numRendered += await handleResultList(results.alternateResults, renderedAlternates, 'Alternate Results');
+  if(numRendered > 0){
+    togglePanel(true);
+  }
+}
+
+const togglePanel = (toggle) => {
+  let container = document.getElementById('IFrame');
+  container.style.display = toggle ? 'block' : 'none';
+}
+
+const hideCard = () => {
+  for(textNode of textNodes){
+    textNode.node.style.visibility = 'hidden';
+  }
+  for(renderedPrimary of renderedPrimaries){
+    mstrHyper.hideCard({ nodeToRenderTo : renderedPrimary.node});
+  }
+  for(renderedAlternate of renderedAlternates){
+    mstrHyper.hideCard({ nodeToRenderTo : renderedAlternate.node});
+  }
+}
+
+const showCard = async () => {
+  for(textNode of textNodes){
+    textNode.node.style.visibility = 'visible';
+  }
+  for(renderedPrimary of renderedPrimaries){
+    await mstrHyper.showCard({elementId: renderedPrimary.result.ref, cardUID: renderedPrimary.result.cardSetId, nodeToRenderTo: renderedPrimary.node});
+  }
+  for(renderedAlternate of renderedAlternates){
+    await mstrHyper.showCard({elementId: renderedAlternate.result.ref, cardUID: renderedAlternate.result.cardSetId, nodeToRenderTo: renderedAlternate.node});
+  }
+}
 
 const deleteHighlights = (e) => {
   const highlights = e.getElementsByTagName('mstr-hi');
@@ -147,14 +278,14 @@ window.addEventListener('DOMContentLoaded', () => {
     .then(() => {
       status(
         `MicroStrategy HyperIntelligence is initialized.<br/><br/>
-You may now open the DevTools and call any of the HyperIntelligence SDK APIs, such as <i>mstrHyper.enableCards</i> and <i>mstrHyper.disableCards</i>.
+You may now open the DevTools and call any of the Hyper SDK APIs, such as <i>mstrHyper.enableCards</i> and <i>mstrHyper.disableCards</i>.
 `,
         'success',
       );
     })
     .catch((error) => {
       status(
-        `Failed to initialize HyperIntelligence SDK, error: \n\n${error.stack}`.replace(
+        `Failed to initialize Hyper SDK, error: \n\n${error.stack}`.replace(
           /\n/g,
           '<br/>',
         ),