+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-# What's new in HyperIntelligence SDK
+# What's new in Hyper SDK
 
-Change log
+## Hyper.Now Jan 2021
+
+- Reuse the MicroStrategy Library Server session if exists.
+- Added `mstrHyper.ssoLogin` API.
+- Added `mstrHyper.searchKeyword` API and related apis to support search.